@@ -21,6 +21,19 @@ from .utils import ChinesePhoneticUtils
 
 
 class ChineseEngine(CorrectorEngine):
+    """
+    中文修正引擎。
+
+    職責：
+    - 持有中文 phonetic backend（pypinyin + 快取）
+    - 建立 phonetic system / tokenizer / fuzzy generator / utils 等共享元件
+    - 將使用者輸入的 term_dict 正規化後，建立輕量的 ChineseCorrector
+
+    設計重點：
+    - backend 單例負責快取與（必要時）依賴檢查
+    - corrector 實例保持輕量，便於快速建立不同 domain 的字典
+    """
+
     _engine_name = "chinese"
 
     def __init__(
@@ -32,6 +45,16 @@ class ChineseEngine(CorrectorEngine):
         verbose: bool = False,
         on_timing: Optional[Callable[[str, float], None]] = None,
     ):
+        """
+        初始化 ChineseEngine。
+
+        Args:
+            phonetic_config: 拼音設定（未提供則使用預設 ChinesePhoneticConfig）
+            enable_surface_variants: 是否自動生成表面變體（包含黏音/懶音等）
+            enable_representative_variants: 是否啟用代表字變體（較昂貴，預設關閉）
+            verbose: 是否輸出較多日誌
+            on_timing: 可選的計時回呼（利於效能觀測）
+        """
         self._init_logger(verbose=verbose, on_timing=on_timing)
 
         with self._log_timing("ChineseEngine.__init__"):
@@ -43,6 +66,7 @@ class ChineseEngine(CorrectorEngine):
             self._tokenizer = ChineseTokenizer()
             self._fuzzy_generator = ChineseFuzzyGenerator(
                 config=self._phonetic_config,
+                backend=self._backend,
                 enable_representative_variants=enable_representative_variants,
             )
             self._utils = ChinesePhoneticUtils(config=self._phonetic_config)
@@ -53,32 +77,40 @@ class ChineseEngine(CorrectorEngine):
 
     @property
     def phonetic(self) -> ChinesePhoneticSystem:
+        """取得中文發音系統（拼音轉換與相似度）。"""
         return self._phonetic
 
     @property
     def tokenizer(self) -> ChineseTokenizer:
+        """取得中文分詞器（用於滑動視窗與 token indices）。"""
         return self._tokenizer
 
     @property
     def fuzzy_generator(self) -> ChineseFuzzyGenerator:
+        """取得中文模糊變體生成器（同音/近音等變體）。"""
         return self._fuzzy_generator
 
     @property
     def utils(self) -> ChinesePhoneticUtils:
+        """取得中文語音工具（聲母/韻母/模糊音判斷等）。"""
         return self._utils
 
     @property
     def config(self) -> ChinesePhoneticConfig:
+        """取得中文語音設定（phonetic config）。"""
         return self._phonetic_config
 
     @property
     def backend(self) -> ChinesePhoneticBackend:
+        """取得中文語音 backend（進階用途：快取統計/清除等）。"""
         return self._backend
 
     def is_initialized(self) -> bool:
+        """檢查 Engine 是否已完成初始化。"""
         return self._initialized
 
     def get_backend_stats(self) -> Dict[str, Any]:
+        """取得 backend 快取統計資訊。"""
         return self._backend.get_cache_stats()
 
     def create_corrector(
@@ -88,6 +120,17 @@ class ChineseEngine(CorrectorEngine):
         on_event: Optional[CorrectionEventHandler] = None,
         **kwargs,
     ) -> ChineseCorrector:
+        """
+        依據 term_dict 建立 ChineseCorrector。
+
+        Args:
+            term_dict: 使用者輸入的字典（支援簡寫格式與完整 config 格式）
+            protected_terms: 不可被替換的保護詞彙（避免誤修）
+            on_event: 可選事件回呼（replacement/degraded/fuzzy_error）
+
+        Returns:
+            ChineseCorrector: 輕量 corrector 實例
+        """
         with self._log_timing("ChineseEngine.create_corrector"):
             normalized_input = normalize_term_dict(term_dict)
 
@@ -100,13 +143,13 @@ class ChineseEngine(CorrectorEngine):
             self._logger.debug(f"Creating corrector with {len(normalized_dict)} terms")
 
             cache_stats = self._backend.get_cache_stats()
-            pinyin_stats = cache_stats.get("pinyin", {})
-            total_hits = pinyin_stats.get("hits", 0)
-            total_misses = pinyin_stats.get("misses", 0)
+            pinyin_stats = cache_stats["caches"].get("pinyin", {"hits": 0, "misses": 0, "currsize": 0, "maxsize": -1})
+            total_hits = pinyin_stats["hits"]
+            total_misses = pinyin_stats["misses"]
             hit_rate = total_hits / max(1, total_hits + total_misses) * 100
             self._logger.debug(
                 f"  [Cache] pinyin: hits={total_hits}, misses={total_misses}, "
-                f"rate={hit_rate:.1f}%, size={pinyin_stats.get('currsize', 0)}"
+                f"rate={hit_rate:.1f}%, size={pinyin_stats['currsize']}"
             )
 
             return ChineseCorrector._from_engine(
@@ -117,6 +160,14 @@ class ChineseEngine(CorrectorEngine):
             )
 
     def _normalize_term_value(self, term: str, value: Any) -> Optional[Dict[str, Any]]:
+        """
+        將 term_dict 的 value 正規化為 internal config dict。
+
+        主要工作：
+        - 統一 aliases / keywords / exclude_when / weight 欄位
+        - 產生 fuzzy variants（可由 enable_surface_variants 控制）
+        - 以拼音去重 aliases，避免字典膨脹（同音不同寫只留一份）
+        """
         if isinstance(value, list):
             value = {"aliases": value}
         elif isinstance(value, dict):
@@ -125,7 +176,7 @@ class ChineseEngine(CorrectorEngine):
         else:
             value = {"aliases": []}
 
-        pinyin = self._utils.get_pinyin_string(term)
+        pinyin = self._backend.to_phonetic(term)
         self._logger.debug(f"  [Pinyin] {term} -> {pinyin}")
 
         merged_aliases = list(value.get("aliases", []))
@@ -150,10 +201,17 @@ class ChineseEngine(CorrectorEngine):
         }
 
     def _filter_aliases_by_pinyin(self, aliases: List[str]) -> List[str]:
+        """
+        依拼音去重 aliases（保留第一個出現的拼寫）。
+
+        目的：
+        - 控制 auto-variants 造成的字典膨脹
+        - 避免同音 alias 重複，讓索引更小、候選生成更快
+        """
         seen_pinyins = set()
         filtered = []
         for alias in aliases:
-            pinyin_str = self._utils.get_pinyin_string(alias)
+            pinyin_str = self._backend.to_phonetic(alias)
             if pinyin_str not in seen_pinyins:
                 filtered.append(alias)
                 seen_pinyins.add(pinyin_str)
