@@ -18,7 +18,8 @@ class DummyEnglishBackend:
 
     def __init__(self):
         self._initialized = True
-        self._stats = {"hits": 0, "misses": 0, "currsize": 0, "maxsize": 999999}
+        self._hits = 0
+        self._misses = 0
 
     def initialize(self) -> None:
         self._initialized = True
@@ -27,10 +28,23 @@ class DummyEnglishBackend:
         return self._initialized
 
     def clear_cache(self) -> None:
-        self._stats = {"hits": 0, "misses": 0, "currsize": 0, "maxsize": 999999}
+        self._hits = 0
+        self._misses = 0
 
     def get_cache_stats(self):
-        return dict(self._stats)
+        # 對齊 BackendStats 統一格式
+        return {
+            "initialized": self._initialized,
+            "lazy_init": {"status": "not_supported"},
+            "caches": {
+                "ipa": {
+                    "hits": self._hits,
+                    "misses": self._misses,
+                    "currsize": 0,
+                    "maxsize": 999999,
+                }
+            },
+        }
 
     @staticmethod
     def _ipa_for(text: str) -> str:
@@ -41,11 +55,11 @@ class DummyEnglishBackend:
         return "paaaa"
 
     def to_phonetic(self, text: str) -> str:
-        self._stats["misses"] += 1
+        self._misses += 1
         return self._ipa_for(text)
 
     def to_phonetic_batch(self, texts: list):
-        self._stats["misses"] += max(0, len(texts))
+        self._misses += max(0, len(texts))
         return {t: self._ipa_for(t) for t in texts}
 
 
@@ -96,21 +110,14 @@ def test_chinese_initials_bucket_prunes_items(monkeypatch):
     """
     from phonofix import ChineseEngine
 
-    engine = ChineseEngine()
-    corrector = engine.create_corrector(["牛奶"])
+    engine = ChineseEngine(enable_surface_variants=False)
 
-    # 關注點：_generate_fuzzy_candidate_drafts 只應迭代相符 group 的 items。
-    # 將 buckets 人工放大：n_l_group 只有 2 個，其他群組有大量假資料。
-    corrector._exact_matcher = None
-    corrector._exact_items_by_alias = {}
+    keep_terms = ["牛奶", "流奶"]  # n / l -> n_l_group
+    noise_prefixes = ["中", "車", "山", "資", "發", "花", "工", "北", "大", "是"]
+    noise_suffixes = ["站", "票", "園", "國", "學"]
+    noise_terms = [p + s for p in noise_prefixes for s in noise_suffixes]
 
-    buckets = {
-        2: {
-            "n_l_group": [{"id": "keep1"}, {"id": "keep2"}],
-            "c_group": [{"id": f"noise{i}"} for i in range(300)],
-        }
-    }
-    corrector._fuzzy_buckets = buckets
+    corrector = engine.create_corrector(keep_terms + noise_terms)
 
     calls = {"n": 0}
 
@@ -118,7 +125,38 @@ def test_chinese_initials_bucket_prunes_items(monkeypatch):
         calls["n"] += 1
         return None
 
-    monkeypatch.setattr(corrector, "_process_fuzzy_match_draft", _counting_process)
+    monkeypatch.setattr("phonofix.languages.chinese.candidates.process_fuzzy_match_draft", _counting_process)
 
-    assert corrector.correct("流奶") == "流奶"
+    # 不走 exact，避免干擾計數；這裡只關注 fuzzy 分桶是否只遍歷同群組的 items
+    corrector._exact_matcher = None
+    corrector._exact_items_by_alias = {}
+
+    assert corrector.correct("流奶", silent=True) == "流奶"
+    assert calls["n"] == 2
+
+
+def test_japanese_group_pruning_limits_similarity_calls(monkeypatch):
+    """
+    日文 fuzzy 產生候選時應以首音群組分桶，避免每個 window 對所有 item 計算相似度。
+    """
+    from phonofix import JapaneseEngine
+
+    engine = JapaneseEngine(enable_surface_variants=False)
+    keep_terms = ["パンダ", "バナナ"]  # p / b -> group 1
+    noise_terms = [
+        "タコ", "トマト", "カメラ", "サケ", "マグロ", "ラーメン", "ヤマ", "ワサビ", "ジャズ", "チーズ",
+        "スシ", "ゼロ", "フジ", "ホン", "ナス", "ネコ", "ゴマ", "ガラス", "レモン", "リンゴ",
+        "イチゴ", "ウドン", "アメ", "オチャ", "キノコ", "テスト", "データ", "ハコ", "ミカン", "ソラ",
+    ]
+    corrector = engine.create_corrector(keep_terms + noise_terms)
+
+    calls = {"n": 0}
+
+    def _counting_similarity(a, b):
+        calls["n"] += 1
+        return 1.0, False
+
+    monkeypatch.setattr(corrector.phonetic, "calculate_similarity_score", _counting_similarity)
+
+    assert corrector.correct("パンダ", silent=True) == "パンダ"
     assert calls["n"] == 2
